backend/requirements.txt
fastapi==0.116.1
starlette==0.46.2
passlib==1.7.4
protobuf==6.31.1
pydantic==2.11.7
pydantic[email]
pydantic_settings
python-multipart==0.0.20 
python-dotenv==1.1.1
python_jose==3.5.0
SQLAlchemy==2.0.42
uvicorn==0.34.2
psycopg2==2.9.10
